+++ src/xtd.forms.native.wxwidgets/src/xtd/forms/native/wxwidgets/tab_page.cpp
@@ -12,7 +12,7 @@ void tab_page::image_index(intptr_t control, size_t image_index) {
   if (control == 0) return;
   auto page = static_cast<wxNotebookPage*>(reinterpret_cast<control_handler*>(control)->control());
   auto note_book = static_cast<wxNotebookBase*>(page->GetParent());
-  if (note_book) note_book->SetPageImage(note_book->FindPage(page), image_index);
+  if (note_book) note_book->SetPageImage(note_book->FindPage(page), static_cast<int>(image_index));
 }
 
 void tab_page::text(intptr_t control, const ustring& text) {
